Created MailConfiguration to provide the conditional beans Moved spring.mail config to example yaml file Null check when checking if email notifications are active Renamed email to mail for consistency
