+++ application.example.yaml
@@ -2,6 +2,19 @@
 # Copy this file to "application.yaml" and fill out the settings that you want for your bot.
 # When you run the JAR file, make sure the config file is in the current directory so it can be read.
 
+# If you wish to be notified by email on every trade then you need to add your mail server configuration here and
+# configure the email notifications in notifications.mail section
+spring:
+  mail:
+    # Your mail server host
+    host: smtp.mail.com
+    username: username@mail.com
+    password: password
+    port: 465
+    properties:
+      # If your email supports SSL or TLS, set this property to true
+      mail.smtp.ssl.enable: true
+
 # All the logs you see on the console will also be shipped to Slack.
 notifications:
   slack:
@@ -15,7 +28,7 @@ notifications:
     # exchange's API is. Because there are so many factors it made sense to be able to configure the warning so you can
     # control how noisy your logs are.
     slowTickerWarning: 4000
-  email:
+  mail:
     # Set to true to enable email notifications for each trade
     active: false
     # The origin email address to set in the email notification
