+++ src/main/resources/application.yaml
@@ -7,12 +7,3 @@ spring:
   jackson:
     serialization:
       write-dates-as-timestamps: false
-  mail:
-    # Your mail server host
-    host: smtp.mail.com
-    username: username@mail.com
-    password: password
-    port: 465
-    properties:
-      # If your email supports SSL or TLS, set this property to true
-      mail.smtp.ssl.enable: true
